+++ .prettierrc.js
@@ -28,6 +28,8 @@ export default {
     "",
     "<TYPES>",
     "",
+    "<TYPES>^[#]",
+    "",
     "<TYPES>^[./]",
   ],
 };
