+++ app/theme/schema.ts
@@ -1,0 +1,5 @@
+import { z } from "zod";
+
+export const ThemeFormSchema = z.object({
+  theme: z.enum(["system", "light", "dark"]),
+});
